Refactor(DocsDto): docs dto 파일 구조 변경
