+++ src/main/java/com/project/bumawiki/domain/docs/presentation/DocsCreateUpdateController.java
@@ -1,9 +1,11 @@
 package com.project.bumawiki.domain.docs.presentation;
 
 
-import com.project.bumawiki.domain.docs.domain.type.DocsType;
-import com.project.bumawiki.domain.docs.exception.DocsTypeNotFoundException;
-import com.project.bumawiki.domain.docs.presentation.dto.*;
+import com.project.bumawiki.domain.docs.presentation.dto.request.DocsCreateRequestDto;
+import com.project.bumawiki.domain.docs.presentation.dto.request.DocsTitleUpdateRequestDto;
+import com.project.bumawiki.domain.docs.presentation.dto.request.DocsTypeUpdateDto;
+import com.project.bumawiki.domain.docs.presentation.dto.request.DocsUpdateRequestDto;
+import com.project.bumawiki.domain.docs.presentation.dto.response.DocsResponseDto;
 import com.project.bumawiki.domain.docs.service.DocsCreateService;
 import com.project.bumawiki.domain.docs.service.DocsUpdateService;
 import lombok.RequiredArgsConstructor;
@@ -28,7 +30,7 @@ public class DocsCreateUpdateController {
     private final DocsCreateService docsCreateService;
 
     @PutMapping("/update/{title}")
-    public DocsResponseDto updateDocs(@RequestHeader("Authorization")String bearer, @PathVariable String title,@RequestPart DocsUpdateRequestDto request,@RequestPart(required = false) MultipartFile[] files) throws IOException {
+    public DocsResponseDto updateDocs(@RequestHeader("Authorization") String bearer, @PathVariable String title, @RequestPart DocsUpdateRequestDto request, @RequestPart(required = false) MultipartFile[] files) throws IOException {
         return docsUpdateService.execute(bearer, title, request, files);
     }
