+++ src/main/java/com/project/bumawiki/domain/docs/presentation/DocsInformationController.java
@@ -3,10 +3,10 @@ package com.project.bumawiki.domain.docs.presentation;
 import com.project.bumawiki.domain.docs.domain.type.DocsType;
 import com.project.bumawiki.domain.docs.exception.DocsTypeNotFoundException;
 
-import com.project.bumawiki.domain.docs.presentation.dto.DocsNameAndEnrollResponseDto;
-import com.project.bumawiki.domain.docs.presentation.dto.DocsResponseDto;
-import com.project.bumawiki.domain.docs.presentation.dto.VersionDocsDiffResponseDto;
-import com.project.bumawiki.domain.docs.presentation.dto.VersionResponseDto;
+import com.project.bumawiki.domain.docs.presentation.dto.response.DocsNameAndEnrollResponseDto;
+import com.project.bumawiki.domain.docs.presentation.dto.response.DocsResponseDto;
+import com.project.bumawiki.domain.docs.presentation.dto.response.VersionDocsDiffResponseDto;
+import com.project.bumawiki.domain.docs.presentation.dto.response.VersionResponseDto;
 import com.project.bumawiki.domain.docs.service.DocsInformationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
