+++ src/main/java/com/project/bumawiki/domain/docs/presentation/dto/request/DocsTitleUpdateRequestDto.java
@@ -1,4 +1,4 @@
-package com.project.bumawiki.domain.docs.presentation.dto;
+package com.project.bumawiki.domain.docs.presentation.dto.request;
 
 import lombok.Getter;
