+++ src/main/java/com/project/bumawiki/domain/docs/presentation/dto/response/VersionDocsResponseDto.java
@@ -1,4 +1,4 @@
-package com.project.bumawiki.domain.docs.presentation.dto;
+package com.project.bumawiki.domain.docs.presentation.dto.response;
 
 import com.project.bumawiki.domain.docs.domain.VersionDocs;
 import com.project.bumawiki.domain.user.entity.User;
