+++ src/main/java/com/project/bumawiki/domain/docs/presentation/dto/response/VersionResponseDto.java
@@ -1,5 +1,7 @@
-package com.project.bumawiki.domain.docs.presentation.dto;
+package com.project.bumawiki.domain.docs.presentation.dto.response;
 
+import com.project.bumawiki.domain.docs.presentation.dto.response.DocsResponseDto;
+import com.project.bumawiki.domain.docs.presentation.dto.response.VersionDocsResponseDto;
 import lombok.Getter;
 
 import java.util.List;
