+++ src/main/java/com/project/bumawiki/domain/docs/service/DocsCreateService.java
@@ -6,8 +6,8 @@ import com.project.bumawiki.domain.docs.domain.Docs;
 import com.project.bumawiki.domain.docs.domain.VersionDocs;
 import com.project.bumawiki.domain.docs.domain.repository.VersionDocsRepository;
 import com.project.bumawiki.domain.docs.facade.DocsFacade;
-import com.project.bumawiki.domain.docs.presentation.dto.DocsCreateRequestDto;
-import com.project.bumawiki.domain.docs.presentation.dto.DocsResponseDto;
+import com.project.bumawiki.domain.docs.presentation.dto.request.DocsCreateRequestDto;
+import com.project.bumawiki.domain.docs.presentation.dto.response.DocsResponseDto;
 import com.project.bumawiki.domain.image.service.ImageService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
