+++ src/main/java/com/project/bumawiki/domain/docs/service/DocsUpdateService.java
@@ -11,10 +11,10 @@ import com.project.bumawiki.domain.docs.exception.CannotChangeYourDocsException;
 import com.project.bumawiki.domain.docs.exception.DocsNotFoundException;
 import com.project.bumawiki.domain.docs.exception.NoUpdatableDocsException;
 import com.project.bumawiki.domain.docs.facade.DocsFacade;
-import com.project.bumawiki.domain.docs.presentation.dto.DocsResponseDto;
-import com.project.bumawiki.domain.docs.presentation.dto.DocsTitleUpdateRequestDto;
-import com.project.bumawiki.domain.docs.presentation.dto.DocsTypeUpdateDto;
-import com.project.bumawiki.domain.docs.presentation.dto.DocsUpdateRequestDto;
+import com.project.bumawiki.domain.docs.presentation.dto.response.DocsResponseDto;
+import com.project.bumawiki.domain.docs.presentation.dto.request.DocsTitleUpdateRequestDto;
+import com.project.bumawiki.domain.docs.presentation.dto.request.DocsTypeUpdateDto;
+import com.project.bumawiki.domain.docs.presentation.dto.request.DocsUpdateRequestDto;
 import com.project.bumawiki.domain.image.service.ImageService;
 import com.project.bumawiki.domain.user.UserFacade;
 import com.project.bumawiki.domain.user.entity.User;
